requirements.txt
Faker==18.9.0
matplotlib==3.7.2
networkx==3.1
python-decouple==3.8
rich==13.5.2
SQLAlchemy==2.0.20
mysql-connector-python==8.1.0